requirements.txt
ics==0.7.2
Jinja2==3.1.2
Pillow==9.5.0
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
selenium==4.9.0
